RavenDB-19100 - Adding validation method for adding node to db
